# ws2026/ModuleA/media/openapi.yaml
openapi: 3.0.3
info:
  title: 'Eventify API Documentation'
  description: 'RESTful API for managing technical events, participants, and registrations.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Token must be a string longer than 32 characters.'
security:
  -
    default: []
paths:
...
  '/api/v1/events/{id}':
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the event.'
        required: true
        schema:
          type: integer
          example: 1

    get:
      summary: 'Event Details'
      operationId: eventDetails
      description: 'Returns detailed information about an event including its confirmed participants.'
      tags:
        - Events
      responses:
        200:
          description: 'Event details'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: string
                      description:
                        type: string
                        example: string
                      location:
                        type: string
                        example: string
                      event_date:
                        type: string
                        example: '2026-06-15'
                      capacity:
                        type: integer
                        example: 100
                      registrations_count:
                        type: integer
                        description: 'Reflects only CONFIRMED registrations.'
                        example: 10
                      participants:
                        type: array
                        description: 'Only participants with CONFIRMED registrations for this event.'
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            name:
                              type: string
                              example: string
                            phone:
                              type: string
                              example: '+77001234567'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized user'
        404:
          description: 'Not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Not found'

    put:
      summary: 'Update Event (ADMIN)'
      operationId: updateEvent
      description: 'Fully updates an existing event. All fields are required. Accessible only for ADMIN. If the capacity is less than the number of confirmed registrations, returns 409.'
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - location
                - event_date
                - capacity
              properties:
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'Tech Conference 2026 Updated'
                description:
                  type: string
                  example: 'Updated description.'
                location:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: Almaty
                event_date:
                  type: string
                  description: 'Format: YYYY-MM-DD. Must be >= today and <= today + 1 year.'
                  example: '2026-07-20'
                capacity:
                  type: integer
                  description: 'Must be at least 1.'
                  example: 150
      responses:
        200:
          description: 'Event updated successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'update success'
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: string
                      description:
                        type: string
                        example: string
                      location:
                        type: string
                        example: string
                      event_date:
                        type: string
                        example: '2026-07-20'
                      capacity:
                        type: integer
                        example: 150
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized user'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Forbidden
        404:
          description: 'Not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Not found'
        409:
          description: 'Conflict — capacity less than confirmed registrations count'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'The capacity must be greater than number of confirmed registrations'
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Validation error'
                  errors:
                    type: object
                    properties:
                      event_date:
                        type: array
                        items:
                          type: string
                        example: ['The event date does not match the format Y-m-d.']

    delete:
      summary: 'Delete Event (ADMIN)'
      operationId: deleteEvent
      description: 'Deletes an event and all related registrations. Accessible only for ADMIN.'
      tags:
        - Events
      responses:
        200:
          description: 'Event deleted successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'delete success'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized user'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Forbidden
        404:
          description: 'Not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Not found'

  # ─── PARTICIPANTS ─────────────────────────────────────────────────────────────
  /api/v1/my-participant:
    get:
      summary: 'My Participant'
      operationId: myParticipant
      description: 'Returns the participant profile belonging to the authenticated user. Returns 404 if the user has not created a profile yet.'
      tags:
        - Participants
      responses:
        200:
          description: 'Participant profile'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: string
                      phone:
                        type: string
                        example: '+77001234567'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized user'
        404:
          description: 'Profile not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Not found'
...
  /api/v1/my-registrations:
    get:
      summary: 'My Registrations'
      operationId: myRegistrations
      description: 'Returns registrations of the authenticated user only.'
      tags:
        - Registrations
      responses:
        200:
          description: 'List of user registrations'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        event:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            title:
                              type: string
                              example: string
                            event_date:
                              type: string
                              example: '2026-06-15'
                        status:
                          type: string
                          enum: [PENDING, CONFIRMED, CANCELLED]
                          example: PENDING
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized user'

  # ─── ADDITIONAL ───────────────────────────────────────────────────────────────
  '/api/v1/events/{id}/my-status':
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the event.'
        required: true
        schema:
          type: integer
          example: 1
    get:
      summary: 'My Status'
      operationId: myStatus
      description: 'Returns the authenticated user registration status for a specific event. Returns null in the data field if the user is not registered.'
      tags:
        - Additional
      responses:
        200:
          description: 'Registration status for the event'
          content:
            application/json:
              schema:
                oneOf:
                  -
                    description: 'User is registered'
                    type: object
                    example:
                      data:
                        event_id: 1
                        status: CONFIRMED
                    properties:
                      data:
                        type: object
                        properties:
                          event_id:
                            type: integer
                            example: 1
                          status:
                            type: string
                            enum: [PENDING, CONFIRMED, CANCELLED]
                            example: CONFIRMED
                  -
                    description: 'User is not registered for this event'
                    type: object
                    example:
                      data: null
                    properties:
                      data:
                        nullable: true
                        example: null
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unauthorized user'
        404:
          description: 'Event not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Not found'